Katalog & info tanaman

+++ app/src/main/java/com/example/authtoga/data/model/Treatment.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class Treatment(
     val id: String? = null,
     // UBAH BARIS INI: Tambahkan tanda tanya (?) setelah String dan berikan nilai default = null
-    val plant_id: Long? = null,
+    val plant_id: Int? = null,
     val title: String,
     val disease_target: String,
     val ingredients: String,
